+++ lib/enums.dart
@@ -1,0 +1,43 @@
+
+enum RecorderAudioChannel { INPUT, OUTPUT }
+
+/// RTCDataChannelMessage type
+enum MessageType { text, binary }
+
+enum RTCDataChannelState {
+  RTCDataChannelConnecting,
+  RTCDataChannelOpen,
+  RTCDataChannelClosing,
+  RTCDataChannelClosed,
+}
+
+enum RTCSignalingState {
+  RTCSignalingStateStable,
+  RTCSignalingStateHaveLocalOffer,
+  RTCSignalingStateHaveRemoteOffer,
+  RTCSignalingStateHaveLocalPrAnswer,
+  RTCSignalingStateHaveRemotePrAnswer,
+  RTCSignalingStateClosed
+}
+
+enum RTCIceGatheringState {
+  RTCIceGatheringStateNew,
+  RTCIceGatheringStateGathering,
+  RTCIceGatheringStateComplete
+}
+
+enum RTCIceConnectionState {
+  RTCIceConnectionStateNew,
+  RTCIceConnectionStateChecking,
+  RTCIceConnectionStateCompleted,
+  RTCIceConnectionStateConnected,
+  RTCIceConnectionStateCount,
+  RTCIceConnectionStateFailed,
+  RTCIceConnectionStateDisconnected,
+  RTCIceConnectionStateClosed,
+}
+
+enum RTCVideoViewObjectFit {
+  RTCVideoViewObjectFitContain,
+  RTCVideoViewObjectFitCover,
+}
