+++ lib/rtc_data_channel.dart
@@ -1,14 +1,8 @@
 import 'dart:async';
-import 'dart:convert';
 import 'dart:typed_data';
 import 'package:flutter/services.dart';
 import 'utils.dart';
-import 'dart:io' show Platform;
-
-
-enum MessageType {
-  text, binary
-}
+import 'enums.dart';
 
 final _typeStringToMessageType = <String, MessageType>{
   'text': MessageType.text,
@@ -74,13 +68,6 @@ class RTCDataChannelMessage {
   /// Use only on binary messages.
   /// See: [isBinary].
   Uint8List get binary => _data;
-}
-
-enum RTCDataChannelState {
-  RTCDataChannelConnecting,
-  RTCDataChannelOpen,
-  RTCDataChannelClosing,
-  RTCDataChannelClosed,
 }
 
 typedef void RTCDataChannelStateCallback(RTCDataChannelState state);
