+++ lib/utils.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/services.dart';
-import 'rtc_peerconnection.dart';
-import 'rtc_data_channel.dart';
+import 'enums.dart';
 
 class WebRTC {
   static const MethodChannel _channel = const MethodChannel('FlutterWebRTC.Method');
